Start preparing for vnode remapping
An Inode object stores its block number, and converts it to a vnode ID
when needed. Requests for the inode ID are changed depending on what
they actually want:

  Internal operations (BPlusTree updating for instance) use the block
  numbers, and interactions with the outside world use the vnode ID.

For now, vnode IDs and block numbers are differentiated by a simple xor
operation in the translation functions.

It's unlikely that all places that need to be changed have been
changed, but things seem to work fine in when testing casually.