Tracker: Add thumbnail support
Define Thumbnail attributes in Attributes.h in such a way that they will
avoid name collisions.
  Media:Thumbnail:16 and 32, 40, 48, 64, etc. for PNG thumbnails
  Media:Thumbnail:CreationTime to see if thumbs need to be regenerated

Allow Tracker to generate thumbnails on read-only volumes, but don't try
to write attributes to the file.

Check the file's modification time in Utilities GetFileIconFromAttr()
and compare it to the thumbnail creation time. If file has not been
modified since the last time we generated thumbnails then return the
thumbnail from the attribute, otherwise fetch a new thumbnail icon
using GetThumbnailIcon().

Add "Generate image thumbnails" Tracker setting. Default is turned off
for now. To generate image thumbnails you must first turn this setting
on in Tracker Windows preferences.

Take the training wheels off -- allow thumbnails to be generated for any
image type that we have a translator for (e.g. SVG, PSD.)

Spawn a get_thumbnail() thread to generate the thumbnail and retreive
it later on from the window thread to fill out the thumbnail icon.
This should improve responsiveness of generating thumbnails from a
folder with a lot of images in it. If on writable BFS volume the
generator thread will write the icon data to an attribute, if not,
it will send the data back to the original thread by calling
send_data(). Calling update_mime_info() instructs the window thread
to come back and look for an icon either in an attribute, or to pick
up the icon data that has been sent through send_data()
using receive_data() filling in the icon data it reads.

Save and retreive PNG image data to attribute instead of using raw
BBitmap data. This trades CPU time to convert the PNGs back and forth
to save disk space used by the thumbnail attributes.

Additional Changes:

Use B_BITMAP_NO_SERVER_LINK on more bitmaps in IconUtils. Look for
HAIKU_TARGET_PLATFORM_HAIKU before using B_BITMAP_NO_SERVER_LINK.

Add std::nothrow to a couple of new bitmap calls in Tracker Utilities
and IconUtils so that we get NULL on failure. Handle NULL in a couple
more cases in IconUtils. Where we don't handle NULL leave the standard
exception throwing new alone.

Move BPrivate::BNodeInfo::CheckNodeIconHintPrivate() to Tracker
Model::CheckNodeIconHint() completing a TODO.

Change-Id: I7f927a5a1f8cf65e4b1aa1e0eb55bbfae87fd969