Tracker: Add thumbnail support
Define thumbnail attributes in Attributes.h:
  Media:Thumbnail to store the thumbnail,
  Media:Thumbnail:CreationTime to see if thumbs need to be regenerated.

Store 128x128 icon in attribute, for icon sizes smaller than 128x128
down-scale the 128x128 icon. Use B_FILTER_BITMAP_BILINEAR to down-scale
the icon using the bilinear scaling algorithm which creates nicer
looking thumbnails than the default scaling algorithm.

Store icons as WebP images which compress smaller than PNGs and fit in
the inode better at 128x128. Thank you Zenja Solaja for the suggestion.
Added B_WEBP_FORMAT to TranslatorFormats.h.

Check the file's modification time in GetFileIconFromAttr() and compare
it to the thumbnail creation time. If file has not been modified since
the last time we generated thumbnails then return the thumbnail from
the attribute, otherwise fetch a new thumbnail with GetThumbnailIcon().

Add "Generate image thumbnails" Tracker setting. Default is turned off
for now. To generate image thumbnails you must first turn this setting
on in Tracker Windows preferences.

Spawn a get_thumbnail() thread to generate thumbnails and retreive them
later on from the window thread to fill out into the icon. This should
improve responsiveness of generating thumbnails from a folder with a lot
of images. The generator thread will write the icon data to an attribute
if on writable BFS volume.

If not on writable BFS volume, the generator thread will send the data
back to the original thread through a port by calling write_port().

When the thread is finished creating the thumbnail it sends a message
back to the Tracker application thread to update the pose which instructs
the window thread to look for an icon. It either finds an icon in an
attribute, or picks up the icon data that has been sent through
write_port() using read_port().

This works on both read-write and read-only BFS volumes but it still
depends on the presence of a BEOS:TYPE parameter to have been written
to the volume before it became read-only. Thumbnail generation does not
work on other read-only volumes for example images on an ISO-9660 CD,
but it does work on read-only BFS volumes for example the BeOS R5 CD.

Move BPrivate::CheckNodeIconHintPrivate() from BNodeInfo to Tracker
Model CheckNodeIconHint(). Create Model::CheckAppIconHint() and look
for a vector icon or mini and large icon in that method. Check that
the base type is directory, volume, trash, desktop, or if executable
call CheckAppIconHint().

Change-Id: I7f927a5a1f8cf65e4b1aa1e0eb55bbfae87fd969