Tracker: Add thumbnail support
Define thumbnail attributes in Attributes.h:
  Media:Thumbnail:16, 32, 64, and 128 to store thumbnails,
  Media:Thumbnail:CreationTime to see if thumbs need to be regenerated.

Store integer multiples of the minimum 16x16 icon size in an attribute.
For other icons smaller than 64x64 (40x40, and 48x48,) down-scale the
64x64 icon. For 96x96 down-scale the 128x128 icon. Generate separate
attributes for 16x16, 32x32, 64x64, and 128x128 icons.

Store icons as PNGs.

Check the file's modification time in Utilities GetThumbnailIcon()
and compare it to the thumbnail creation time. If file has not been
modified since the last time we generated thumbnails then return the
thumbnail from the attribute, otherwise fetch a new thumbnail icon.

Add "Generate image thumbnails" Tracker setting. Default is turned off
for now. To generate image thumbnails you must first turn this setting
on in Tracker Windows preferences.

Spawn a get_thumbnail() thread to generate thumbnails and retreive
them later on from the window thread to fill out into the icon.
This should improve responsiveness of generating thumbnails from a
folder with a lot of images in it. If on writable BFS volume the
generator thread will write the icon data to an attribute

If not on writable BFS volume, the generator thread will send the data
back to the original thread through a port by calling write_port().

When the thread is finished creating the thumbnail it sends a message
back to the Tracker application thread to update the pose icon which
instructs the window thread to look for an icon. It either finds an
icon in an attribute, or picks up the icon data that has been sent
through write_port() using read_port().

This works on both read-write and read-only BFS volumes but it still
depends on the presence of a BEOS:TYPE parameter to have been written
to the volume before it became read-only. Thumbnail generation does not
work on other read-only volumes like for example images on an ISO-9660
CD, but it does work on read-only BFS volumes, including those from CDs.

Take the training wheels off -- allow thumbnails to be generated for any
image type that we have a translator for (e.g. SVG, PSD.)

Additional Changes:

Use B_BITMAP_NO_SERVER_LINK on more bitmaps in IconUtils. Look for
HAIKU_TARGET_PLATFORM_HAIKU before using B_BITMAP_NO_SERVER_LINK.

Add std::nothrow to a couple of new bitmap calls in Tracker Utilities
and IconUtils so that we get NULL on failure. Handle NULL in a couple
more cases in IconUtils. Where we don't handle NULL leave the standard
exception throwing new alone.

Move BPrivate::CheckNodeIconHintPrivate() from BNodeInfo to Tracker
Model CheckNodeIconHint(). Create Model::CheckAppIconHint() and look
for a vector icon or mini and large icon in that method. Check that
the base type is directory, volume, trash, desktop, or if executable
check CheckAppIconHint().

Change-Id: I7f927a5a1f8cf65e4b1aa1e0eb55bbfae87fd969