Tracker: Add thumbnail support
Define thumbnail attributes in Attributes.h:
  Media:Thumbnail to store the thumbnail,
  Media:Thumbnail:CreationTime to see if thumbs need to be regenerated.

Store 128x128 icon in attribute, for icon sizes smaller than 128x128
down-scale the 128x128 icon. Use B_FILTER_BITMAP_BILINEAR to down-scale
the icon using the bilinear scaling algorithm which creates nicer
looking icons than the default scaling algorithm. Store icons aso PNGs.

Check the file's modification time in Utilities GetThumbnailIcon()
and compare it to the thumbnail creation time. If file has not been
modified since the last time we generated thumbnails then return the
thumbnail from the attribute, otherwise fetch a new thumbnail icon.

Add "Generate image thumbnails" Tracker setting. Default is turned off
for now. To generate image thumbnails you must first turn this setting
on in Tracker Windows preferences.

Spawn a get_thumbnail() thread to generate thumbnails and retreive
them later on from the window thread to fill out into the icon.
This should improve responsiveness of generating thumbnails from a
folder with a lot of images in it. If on writable BFS volume the
generator thread will write the icon data to an attribute.

Do an InitCheck() on the thumbnail params before spawing the thread. In
the unlikely event that we don't have enough memory available to copy
the thread params to be used by the thumbnail generation thread we exit
with an error instead of spawning a thread that cannot succeed at
generating a thumbnail because the system doesn't have enough memory.

If not on writable BFS volume, the generator thread will send the data
back to the original thread through a port by calling write_port(),
otherwise we write the thumbnail data to an attribute.

When the thread is finished creating the thumbnail it sends a message
back to the Tracker application thread to update the pose which
instructs the window thread to look for an icon. It either finds an icon
in an attribute, or picks up the icon data that has been sent through
write_port() using read_port().

This works on both read-write and read-only BFS volumes but it still
depends on the presence of a BEOS:TYPE parameter to have been written
to the volume before it became read-only. Thumbnail generation does not
work on other read-only volumes for example images on an ISO-9660
CD, but it does work on read-only BFS volumes, including those on CDs.

Take the training wheels off -- allow thumbnails to be generated for any
image type that we have a translator for (e.g. SVG, PSD.)

Additional Changes:

Rename mimeType to type in Tracker Utilities and IconUtils.

Use B_BITMAP_NO_SERVER_LINK on more bitmaps in IconUtils. Remove
HAIKU_TARGET_PLATFORM_HAIKU check as we assume that IconUtils is
Haiku only.

Add std::nothrow to a couple of new bitmap calls in Tracker Utilities
and IconUtils so that we get NULL on failure. Handle NULL in a couple
more cases in IconUtils. Where we don't handle NULL leave the standard
exception throwing new alone.

Move BPrivate::CheckNodeIconHintPrivate() from BNodeInfo to Tracker
Model CheckNodeIconHint(). Create Model::CheckAppIconHint() and look
for a vector icon or mini and large icon in that method. Check that
the base type is directory, volume, trash, desktop, or if executable
check CheckAppIconHint().

Change-Id: I7f927a5a1f8cf65e4b1aa1e0eb55bbfae87fd969