Tracker: Add thumbnail support
Define thumbnail attributes in Attributes.h:
  Media:Thumbnail:64, 96, and 128 to store thumbnails,
  Media:Thumbnail:CreationTime to see if thumbs need to be regenerated.

Store a minimum icon size of 64x64 in an attribute. For smaller icons
(16x16, 32x32, 40x40, and 48x48) down-scale the 64x64 icon. Generate
separate attributes for 96x96 and 128x128 icons. Store icons as PNGs.

Check the file's modification time in Utilities GetFileIconFromAttr()
and compare it to the thumbnail creation time. If file has not been
modified since the last time we generated thumbnails then return the
thumbnail from the attribute, otherwise fetch a new thumbnail icon
using GetThumbnailIcon().

Add "Generate image thumbnails" Tracker setting. Default is turned off
for now. To generate image thumbnails you must first turn this setting
on in Tracker Windows preferences.

Spawn a get_thumbnail() thread to generate thumbnails and retreive
them later on from the window thread to fill out into the icon.
This should improve responsiveness of generating thumbnails from a
folder with a lot of images in it. If on writable BFS volume the
generator thread will write the icon data to an attribute, if not,
it will send the data back to the original thread by calling send_data().

When the thread is finished creating the thumbnail it calls
update_mime_info() to instruct the window thread to come back and look
for an icon. It either finds an icon in an attribute, or picks up the
icon data that has been sent through send_data() using receive_data().

Allow Tracker to generate thumbnails on read-only volumes, but don't try
to write attributes to the file.

Take the training wheels off -- allow thumbnails to be generated for any
image type that we have a translator for (e.g. SVG, PSD.)

Additional Changes:

Use B_BITMAP_NO_SERVER_LINK on more bitmaps in IconUtils. Look for
HAIKU_TARGET_PLATFORM_HAIKU before using B_BITMAP_NO_SERVER_LINK.

Add std::nothrow to a couple of new bitmap calls in Tracker Utilities
and IconUtils so that we get NULL on failure. Handle NULL in a couple
more cases in IconUtils. Where we don't handle NULL leave the standard
exception throwing new alone.

Move BPrivate::CheckNodeIconHintPrivate() from BNodeInfo to Tracker
BPrivate::Model CheckNodeIconHint().

Change-Id: I7f927a5a1f8cf65e4b1aa1e0eb55bbfae87fd969