kernel: Add event queue implementation to wait for objects efficiently.
Adds a kqueue/epoll like interface for monitoring file descriptors and
other objects.

event_queue_create() creates a queue and returns an fd for the queue
event_queue_select() adds, removes and modifies monitored objects
event_queue_wait() waits for events on the queue

Events are currently edge-triggered only. When an object is deleted,
a B_EVENT_INVALID event is delivered, and the object is unregistered
from the queue.

The special event flag B_EVENT_ONE_SHOT can be passed in when adding
an object so that the object is automatically unregistered when an
event is delivered.

Modifications-by: Augustin Cavalier <waddlesplash@gmail.com>

Removed public interface (syscalls remain private for the moment),
removed select/modify/delete flags (select/modify are now the same
operation, and "delete" is done when 0 is passed for "events"),
fixed reference counting, memory usage in syscalls, deduplified
code with wait_for_objects, and rewrote event dequeueing logic.

Change-Id: I1d2f094fd981c95215a59adbc087523c7bbbe40b