App Server: Override BeOS app font to 10pt SwisBT 720
... and 12pt SwisBT 720 bold for bold font and 12pt Courier10 BT
for fixed font.

Fall back to Noto Sans Display Regular if SwiBT 720 is not found for
plain and bold fonts, or Noto Sans Mono (not Noto Sans) Regular if
Courier10 BT is not found for fixed font. Haiku apps fall back to
Noto Sans Thai (and not Noto Sans CJK JP) Regular for Haiku apps as
that is what we're doing in GlyphLayoutEngine.h.

Put common code in a private _SetDefaultFonts() method which sets
fPlainFont, fBoldFont, fFixedFont variables. This is called from
AS_GET_SYSTEM_FONTS.

Change-Id: Ib6e137a74a086fb16f0c3c8b372211a139c00017