App Server: Detect BeOS app and override font to 10pt SwisBT 720
and 12pt SwisBT 720 bold for bold font and 12pt Courier10 BT for fixed font.

Fall back to Noto Sans Display or Noto Mono if SwiBT 720 or Courier10 BT is
not found. Fall back to Noto Sans CJK JP for Haiku apps as that is what
we're already doing in GlyphLayoutEngine.h.

Put common code in a private _SetDefaultFonts() method which sets fPlainFont,
fBoldFont, fFixedFont variables. This is called from AS_GET_SYSTEM_FONTS.