TextView: Reimplement the behavior of ScrollToOffset function..
According to the document: "Scroll the text so that the character at offset is visible", this is
not true in case of deleting characters when texts grow more than the bounding view with text
alignment B_ALIGN_RIGHT or B_ALIGN_CENTER set.
The new implementation changes the position of text (at offset) when it grows more than the bounding
view, so it is either at the start or last of the view.