Implement new `setlocale' command
This is a command line tool that serves the same purpose as the "Locale"
application under preferences. It allows the user to set the system
locale with syntax similar to the following:

  setlocale en
  setlocale en us -v posix
  setlocale bs -s cyrl
  setlocale bs ba -s cyrl

It also supports the -l flag to list all available locales.

In the future, the setlocale utility should support all the features
available from the GUI application.