Bind file locks to file descriptors, rather than teams.
File locks created by flock should only apply for the file descriptor
that was used to lock the file. Another fd on the same file should then
be denied access (calling flock should fail).

fcntl based locks, however, are in a separate namespace and are global
to a team.

This issue was found when running webkitpy test suite.

ticket: #13795
Change-Id: I9efb96cfcefe7e72b0060220c635a665e7e643cc