Bind file locks to file descriptors, rather than teams.
File locks created by flock should only apply for the file descriptor
that was used to lock the file. Another fd on the same file should then
be denied access (calling flock should fail).

fcntl based locks, however, are in a separate namespace and are global
to a team.

Note that this patch is incomplete: when closing a file, all associated
locks should be revoked, but we don't get access to the file descriptor
at that point so we can't locate them (see FIXME in the code).

This issue was found when running webkitpy test suite.

ticket: #13795
Change-Id: I9efb96cfcefe7e72b0060220c635a665e7e643cc