BView: Hack a pen location update to fix printing.
When recording a BPicture, the app_server just record events and does
not draw anything. As a result, drawing code that relies on the view
state can get out of sync.

Specifically, our text drawing code relies on getting the pen position
from app_server when asked to draw at "current pen position". As a
result, the pen is never moved and when drawing multiple chunks of text,
they end up overlapping, all drawn at the same cursor position.

To avoid the problem and show that this is indeed what is happening,
this patch adds an explicit cursor move into the BPicture after drawing
a string. This also updates app_server internal pen position, so further
drawing happens at the expected place.

This is however not the right way to fix this. Either we should remove
the code that gets the pen position back from app_server, and let
app_server completely manage the cursor position, or, we should make
sure recording a BPicture also updates the pen position so the
application can rely on accessing it and getting up to date data.

Fixes #6761.

NOTE(waddlesplash): It seems to make sense to include this hack
in R1/beta1, so that printing is not totally broken. Obviously
it should stay out of master.

Change-Id: Ib8e031355032228ef29aceee5f1e1ca21565b831