Update patch set 1
Patch Set 1: Code-Review+2

This structure is mapped directly on the on-disk data, so it must match the on-disk data. This means we can't have a field for the attribute, it is just stored right after the name. And the access as it is done here (at name+name_length) is right.

We should merge that because the current code (pretending name is a pointer) is wrong and dangerous (it will take the first 4 chars of the name as a pointer).

Patch-set: 1
Label: Code-Review=+2