Update patch set 1
Patch Set 1: Code-Review+2

To me this changeset still seems useful. While we might want a grander solution in the future, it is unclear when/if someone's going to implement it. When looking towards a release soon, making our package selection more visible is important to show users that we have, in fact, a lot of software available.

The current HaikuDepot UI conceals this, as the "show featured" option is hidden away in a menu. Many users will think that the featured packages are all there is. That entering something in the search box always unlocks the full package list is quite obscure too - if I didn't know it already, I wouldn't expect it. Making "show recommended" a checkbox in a prominent place of the main window makes it more discoverable that there's more packages to be found.

So at least as a stopgap solution, I support applying this, since we want to put out a beta soon. Unless someone here volunteers to work on a larger remodel of the main window...

Patch-set: 1
Reviewer: Gerrit User 1000028 <1000028@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+2