Update patch set 1
Patch Set 1: Code-Review-1

I already suggested an alternative which is not a very big change: just make the search bar bigger to make it obvious it is the main widget and the way you are supposed to interact with the application.

I think the main problem in the UI is that currently it looks like the search bar is a side-widget of the main view, whereas it is not the case (using the search bar brings you to a new view on the packages).

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-1