Update patch set 1
Patch Set 1:

Making the search bar bigger isn't enough IMO. There also needs to be a clearer visual indication that the "recommended package" list is just that: a *selection* of what's there in total.

You posted a Play Store screenshot, which conveys this clearly: there's headers "New & Updated", "Recommended for You", etc. The list view in Haiku however makes it look like it's all there is.

(Note also what Stephan wrote in the trac ticket: the way it currently looks was not supposed to be the final form; what he describes would certainly much improve it).

Patch-set: 1