Update patch set 1
Patch Set 1:

On a further note, I think that our search functionality is currently not nearly good enough to make it useful for "search is the way to discover what's there".
Just as an example. Suppose I'm a new user and I want to find graphics software. I enter "graphics" in the search box. What I get is a list of almost only library packages with cryptic names. However, it does not show ArtPaint, not BeCasso, not Album, none of those. Ok, let's try searching for "image editor". What I get is 2 packages: gutenprint and nanosvg-translator. Again, none of our actual image editors show up.

Making the search function good enough for that kind of thing might be quite difficult. Even the large players like Apple struggle with it, the search function in their Appstore often returns lots of unexpected results.

So I'd rather focus the UI on the category selection for discovery for now, and leave the search bar the small thing in the corner, for the cases where someone knows exactly (by name) what they're looking for.

Patch-set: 1