Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (2 comments)

The assumption is that if a font has fixed-with ASCII characters, that the rest of the set will either be the same width or double-wide. Terminal already uses a check nearly identical to this one to determine if a font is suitable for its usage.

FontConfig iterates over all available characters, which we could do as well.

Patch-set: 1