Update patch set 1
Patch Set 1:

> (1 comment)

I understood the problem...

The doubt I'm having now is
_QuitApps uses an app list which is created before it is called.
And it's the one which sends msgs to apps that system is shutting down

Now If I set ShuttingDownPhase in early in _QuitApps, it has the same effect as before.
If I set it late in _QuitApps, there might be some apps newly created not in the list.

I'll think more on this problem
If you have any intuition, plz let me know

Patch-set: 1