Update patch set 2
Patch Set 2:

From the coding guidelines:

"Use full names such as message over abbreviations such as msg. Use rect, frame, bounds over r, menuItem over mi."

As for the ownership, I don't know the code well enough to give a direct solution. If you make the object a BHandler and attach it to a BLooper, as I said the object will be automatically destroyed when the BLooper (the window, in this case) is destroyed. Is that acceptable for playlist items? To know, we will need to search the existing lifecycle. When are they referenced? By which part of the code? Are there case where we need them to stay live when the window is closed?

According to the results of this investigation we can decide on a new way to manage the lifecycle of these objects. Maybe making them BHandlers automatically destroyed with the window is ok, or maybe not. There are many directions we can go depending on this. Maybe a BHandler attached to the window is fine after all. Maybe we should attach the BHandler to the application instead. Maybe we need to create a BHandler independantly from the playlist item itself everytime the playlist window is opened or closed.

Patch-set: 2