TextView: Fix text don't align when inserting or removing..
BTextView::ScrollToOffset now scrolls respect to the alignment. The idea here
is shifting locally the rect view or textrect based on the distance changed,
an extra argument is added to handle this. The distance is obtained by calculated
the difference after and before changing text.

There are some issues follow-up:
  +  Text with B_ALIGN_CENTER won't align nicely because the current version of
     BView::ScrollBy doesn't support fractional scroll.
  +  The distance changed is horizontal difference for now because I haven't tested
     with text with many lines (Will fix if there is problem reported).
  +  The function ScrollToOffset should be renamed as it also shifts the text rect
     (haven't think of any suitable names yet..)

Refixes #13796