HaikuDepot : Coalesce Loading Change Notifications
When the application is starting, it will load in bulk data.  As it
does so, it may issue some change notifications to listeners.  For
each small change in packages' data, there is a notification.  The
system was built to accommodate coalesced changes.  This commit will
coalesce those changes during the bulk load for cases where there
are listeners present.