Update patch set 1
Patch Set 1:

> Sorting fSystemApps is only needed one time after the user phase is ended.
Made the changes

 > Generally speaking, the content of the while loop looks like the block before phase 1. The block before phase 1 could be removed, then the "while" loop could be changed in a "do while" loop.

"do while" loop doesn't work in this scenario. This is because, _QuitApps requires list of user apps as well as it modifies (removes exited apps) the user apps list.

if we write do while loop like this:
 do:
   get fUserApps
   _QuitApps(fUserApps)


Due to this we need additional get of user apps

Patch-set: 1
Label: Code-Review=0