Update patch set 2
Patch Set 2:

> Sorting fSystemApps is only needed one time after the user phase is ended.

Made the required changes.


> Generally speaking, the content of the while loop looks like the block before phase 1. The block before phase 1 could be removed, then the "while" loop could be changed in a "do while" loop.


"do while" loop doesn't work in this scenario. This is because, _QuitApps requires list of user apps as well as it modifies (removes exited apps from) the user apps list.

if we write do while loop like this:
 do:
   get fUserApps
   _QuitApps(fUserApps)
 while fUserApps is not empty

After the call to _QuitApps(), fUserApps is emptied.
Now the loop won't iterate further, but there might be new user apps created.
Due to this we need additional get of user apps.

> Please add a blank after "while" and before the brace

Made this change too...


PS: Ignore the last comment, it was an early send.

Patch-set: 2
Label: Code-Review=0