Update patch set 1
Patch Set 1:

I'm bored of people not looking at the real situation of the media_kit.
The "just play that damn sound" API is simple as it follow:

* Create a BMediaFile
* Pass it to BSoundPlayer
* BSoundPlayer->Play()

As said the reason for having all those classes in places is that they were introduced at different stages of BeOS development and were hard to remove when R5 was created. In fact, look at those classes I mentioned, they are all duplicate functionality.

Patch-set: 1