Update patch set 1
Patch Set 1:

> I'm bored of people not looking at the real situation of the media_kit.
> The "just play that damn sound" API is simple as it follow:
> 
> * Create a BMediaFile
> * Pass it to BSoundPlayer
> * BSoundPlayer->Play()

Ok, that sounds almost simple enough.

Maybe it would be worth your time to write some pages for the Haiku Book, because the Media Kit docs there are quite unhelpful:

https://www.haiku-os.org/docs/api/group__media.html

And even the Be Book is unhelpful because it fails to document BSoundPlayer (copypaste error, I guess?):
https://www.haiku-os.org/legacy-docs/bebook/BSoundPlayer.html

So I will do my homework and look at the header file:
https://git.haiku-os.org/haiku/tree/headers/os/media/SoundPlayer.h

I don't see a way to plug a BMediaFile in there. So please give me a complete example, because your "just pass it to BSoundPlayer" doesn't look right. I would be happy to update the docs and update Sounds preferences to be an example of the good practise then.

> As said the reason for having all those classes in places is that they were introduced at different stages of BeOS development and were hard to remove when R5 was created. In fact, look at those classes I mentioned, they are all duplicate functionality.

Some of these (the ones in the Game Kit) I think tried to give lower level access to the hardware. But that hardly matters to us as we fake them using the media kit anyways (the situation is similar for BDirectWindow and BWindowScreen with app_server).

Patch-set: 1