Update patch set 1
Patch Set 1:

Discussing this stuff here is very uncomfortable BTW.

I wasn't probably clear enough my bad. I meant that this feature should be added to BSoundPlayer and the other classes should be deprecated.

In BeOS R4 there wasn't BGameSound and BMediaFile. They supplied only BSound which was mostly used with hardcoded wav streams.

BMediaFile wasn't present because at the time they thought they could use BMediaNode/BFileInterface for handling codecs, they soon realized it was too slow and counterproductive to do so. So, the codec API was introduced and BMediaFile is built upon it.

BSoundFile is a relic of the BSubscriber API. It was hidden from BeOS R4 documentation and there should be a reason for that.

BGameSound was added because they wanted an API which is independent from the media_kit as they probably knew at this point they would have deprecated the media_kit. They just wanted to give games long ABI support.

Patch-set: 1