Update patch set 1
Patch Set 1:

Already has +2 from Julian and myself; Humdinger created the patch and so approves, Stephan approved in the ticket, and Janus seems "OK" with it although wants further changes. Submitting even despite Adrien's -1; if it really turns out as annoying as he finds it, we can either revert or improve.

Patch-set: 1