Update patch set 2
Patch Set 2: Code-Review+2

If I understand correctly, we are already building different gensyscalls for each target architecture, but we were accidentally using the same .o files.

This was only working because gcc2 and gcc5 have the same types/sizes in our case so the two versions would behave the same, but we were goign to hit a problem if we tried to run Haiku with a 32/64bit hybrid userland, for example.

Is that right?

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+2