Update patch set 2
Patch Set 2: Code-Review-1

Uploaded a new patch which changes the behavior only for layouted views.

I'm still not entirely happy with it though, because if tab is not attached to any tab view, it's still going to change child view's name.

Adrien might have been right on the ML that there is no way to fix this without breaking ABI :/

Patch-set: 2