Update patch set 1
Patch Set 1:

The scoring mechanism is pretty much the same currently used in Tracker to identify volumes. However, that one actually uses `BVolumes`, so I could not use that here.

If the partition code would offer UUIDs, one could easily add it to the scoring mechanism (which would be great, as the strongest score on Tracker could not be used; the creation time is only available once the volume is mounted).

Patch-set: 1
Label: Code-Review=0