Update patch set 1
Patch Set 1:

Yes, checking over all CJK characters to make sure not a single one is out of range would be the completely correct way to do this, but it may be quite expensive (that's a lot of characters to check).

Maybe we can check one or two to see if they are indeed double the size of ASCII chars. Which characters should be used as reference points here? Which unicode pages should be checked? Just CJK or are there other alphabets needing double width chars?

Patch-set: 1