Update patch set 6
Patch Set 6:

> Patch Set 6: Code-Review+2
> 
> (1 comment)

Just to add to my comment. The session (or PID) is only relevant for POSIX locks, as they are indeed are based on a vnode-pid pair. Also here I am questioning whether a killed process cleans up properly. Here you will have the same issue that the cleanup is done from the kernel (session 1). This might cause problems when the pid is reused by a process that uses the exact same locking on the exact same files.

Patch-set: 6
Label: Code-Review=0