Update patch set 7
Patch Set 7:

Well, the way it works is that release_advisory_lock releases either locks bound to a file_descriptor, or, in the special case of being passed a NULL descriptor, it releases fcntl-type locks, which are bound to a team id instead - and this raises the question of wether such locks should actually be released when closing the file descriptor, or only explicitly when the process requests so?

I will see if it's possible to split apart the two types of locks to make things clearer.

Patch-set: 7
Label: Code-Review=0