Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> Please include more details about why these are needed in the commit message. I have an idea for the two others, but I'm particularly curious about virtio_rng.

Oops, I don't really have a valid reason for including virtio_rng. I included it because otherwise it would be the only virtio-related stuff to not be in HaikuBootstrap.
(Well, I thought that it could be usefull if we ever choose to use a secure communication in port 4242.)
Shall I remove virtio_rng? (Or shall we consider that its uncompressed size, 14 KiB, is small enough for it to be included without a valid reason?)

Patch-set: 1
Label: Code-Review=0