Update patch set 4
Patch Set 4:

> Patch Set 4:

I can add back (to this patch) the virtio_rng bus but I would not know what to tell about it, except that it could be useful someday if we ever choose to add support for encryption in the transfers between bootstrap_daemon.py and bootstrap_client.py.

Should I add back the virtio_rng bus? Thanks!

Patch-set: 4