Update patch set 13
Patch Set 13:

(6 comments)

Kacper, resolved all your concerns.

Humdinger, I added the extra blank line between functions. In addition I restored the B_TRANSLATE to using the << operator, as I tested my change using the ReplaceFirst technique you described earlier and it did not work correctly when switching to a different Locale. The translation has an extra space at the end that may be causing the issue. To properly resolve this I believe I would need to add a %path variable to all the existing translations catalogs, and that feels beyond the scope of the change I'm trying to make.

Patch-set: 13
Label: Code-Review=0