freebsd11_network: Implement counter(9) using atomics.
FreeBSD's method is different: it has a per-CPU data storage structure
that it adds to whenever someone calls add(), and then only locks and
coalesces it whenever fetch() is called. This means that on some
architectures (e.g. x86_64), adding to the counter is one instruction.

However, this seems to be for the most part overengineering, as its
only uses seem to be statistical counting in semi-performance-critical paths.
Axel noted in #12328 that there's a potential way to implement FreeBSD's
method on Haiku using cpu_ent, but that atomics were "perfectly fine",
so we will go with that for now.