Update patch set 4
Patch Set 4:

I don't think that's a huge concern, tbh. And like I said, the `vmpkg` url will go away, so that's the major reason to not use it.

Oh, and for the `Haiku` repositories, the canonical URL for those is actually at `download.haiku-os.org`. (The links in the file now are just redirects.) Those files are actually used.

Patch-set: 4