BeDecorator: update close and zoom button gradients
... to better match BeOS R5

In BeDecorator up until this point we have used a perfectly smooth
linear gradient that was nicely provided by Stippi to draw the close
and zoom buttons, but, BeOS used a more styalized gradient instead.
This commit seeks to reproduce the styalized BeOS gradient on
BeDecorator adjusting itself for larger button sizes.

Consolidate and correct colors. button, light1, light2, and
shadow1 colors are the same on close and zoom buttons. Zoom has
a lighter shadow2 than close button. Try to get as close to the
colors used on BeOS R5 as possible while still allowing the user
to adjust the base color. Added comments indicating the colors used
on BeOS R5 in the active an inactive states for future reference.

Update close and zoom button down states as well. The position of the
elements is rotated by 180 degrees so that the light parts are drawn in
the top left and the dark parts are draw in the bottom right. In the
case of the big square of the zoom button some parts are obscured
by the small square so we don't draw those.

Change-Id: Ic81837aa387f05b04dda3e1ff76fdf103c93f8e8