Update patch set 4
Patch Set 4:

Changing the existing repositories' repo.info's should happen for the Haiku repositories as soon as the next repo rebuild after this change occurs; and for the HaikuPorts repositories as soon as vmpkg is deprecated.

The canonical URLs for the Haiku repositories you can obtain from the redirects. For the HaikuPorts repositories, it's the 'eu.hpkg' one.

Patch-set: 4
Label: Code-Review=0